feat: create employees module

+++ backend/utils/list.go
@@ -1,6 +1,6 @@
 package utils
 
 type List struct {
-	Data  any `json:"data"`
-	Count int `json:"count"`
+	Data  any   `json:"data"`
+	Count int64 `json:"count"`
 }
